- 금액 변환 컴포넌트
