+++ Front/project_Full/src/js/pages/Travel/TravelPackList.js
@@ -2,7 +2,7 @@ import { DataGrid } from '@mui/x-data-grid';
 import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
-import { SERVER_URL } from 'js';
+import { SERVER_URL, ToggleCell } from 'js';
 import { useCheckLogin } from 'js/useCheckLogin';
 
 import LuggageIcon from '@mui/icons-material/Luggage';
@@ -82,35 +82,7 @@ function TravelPackList() {
         );
     }
 
-    /* 금액 표시 */
-    /*▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤*/
-    const formatPrice = (price) => {
-        //가격을 만원, 천원으로 분리
-        const unit = price >= 10000 ? '만' : '';
-        const mainPrice = Math.floor(price / (unit === '만' ? 10000 : 1000)); //만 단위로 표시하면 만 단위의 가격을 계산, 그 외에는 천 단위로 계산
-        const remainder = price % (unit === '만' ? 10000 : 1000); //remainder: 만 단위로 표시되면 가격을 1만으로 나눈 나머지를, 그렇지 않으면 1천으로 나눈 나머지를 계산
-
-        // 포맷된 문자열 생성
-        const formattedPrice = `${mainPrice}${unit}${remainder > 0 ? ` ${remainder}` : ''}원`; //가격 문자열
-
-        return formattedPrice;
-    };
-
-    /*datagrid의 행의 금액에 대한 '금액'과 '한국 통화 형식'변환*/
-    const ToggleCell = ({ value }) => {
-        const [toggle, setToggle] = useState(false);
-
-        const handleClick = () => {
-            setToggle(!toggle);
-        };
-
-        return (
-            <div onClick={handleClick} style={{ cursor: 'pointer' }}>
-                {toggle ? value.toLocaleString() + `원` : formatPrice(value)}
-            </div>
-        );
-    };
-    /*▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤▤*/
+    
 
     /* <패키지여행> */
     /* ----------------------------------------------------------------------- */
